shutter: added up/down and stop support
Added oids for up/down (true/false) and stop (true = stop).
Send on these oids, if present, when user clicks on "big button".
Second click on "same" button will send "stop".
To enable this, a variable "stopMode" is set when sending up/down.
"stopMode" should get disabled when blinds stopped moving. Not sure
how to ensure that. My blinds do send the current position when they are
done, but this is probably not true for all blinds.
Working-ID would be one thing (but I did not yet understand the code
for this and also have no working id for testing).
Another possibility would  be to add another parameter for a timeout, ie
afte pressing up/down the button will issue stop for x seconds, after
that
the button will become up/down again.

Another possibility would be to add a dedicated stop button if there is
a
stop oid. But I lack CSS (and design) knowledge to implement this.

Signed-off-by: Achim Königs <garfonso@mobo.info>